requirements.txt
absl-py==0.9.0
antlr4-python3-runtime==4.8
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
future==0.18.2
google-auth==1.17.2
google-auth-oauthlib==0.4.1
grpcio==1.29.0
hydra-core==1.0.0rc3
idna==2.9
importlib-metadata==1.6.1
importlib-resources==3.0.0
joblib==0.15.1
Markdown==3.2.2
numpy==1.18.5
oauthlib==3.1.0
omegaconf==2.0.1rc12
pandas==1.0.4
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytorch-lightning==0.7.6
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
threadpoolctl==2.1.0
torch==1.5.0
torchaudio==0.5.0
tqdm==4.46.1
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
